Change logs 1. Implementing address swizzling via cute::Swizzle from CuTe library as an alternative to avoid bank-conflict of shared memory. Note that, however, my test has shown a slight slowdown when swizzling method is substituted for padding, and therefore this feature is disabled by default, but could be enabled by the compile flag BANK_CONFLICT_FREE_BY_SWIZZLING.
2. Switching from int64 to int32 integers whenever applicable.
Originally, several integer variables, especially those
representing dimensions, were declared as int64, which incurred a
number of int64 operations per kernel launch. This change reduces
those overheads, giving an extra boost to the execution speed.

+++ flash_attention/kernel/sync_methods.h
@@ -9,7 +9,9 @@ class TensorShape;
 }
 
 struct SequenceDescriptor {
-  using Vector = std::vector<int64_t>;
+  // Note that while int64_t is used to store the number of dimensions,
+  // currently the sizes of each dimension, and therefore indices of them, are limited to the range representable by int32_t
+  using Vector = std::vector<int32_t>;
   SequenceDescriptor(const int64_t dims) {
     shape.reserve(dims);
     stride.reserve(dims);
@@ -40,7 +42,7 @@ struct SequenceDescriptorPack {
 
 using MethodImplPtr = SequenceDescriptorPack (*)(const tensorflow::TensorShape&, const tensorflow::TensorShape&);
 
-template <int SequenceDims>
+template <int32_t SequenceDims>
 class SyncMethod {
  public:
   SyncMethod(const MethodImplPtr method_impl_ptr)
@@ -91,7 +93,7 @@ class SyncMethods {
   SyncMethods() = delete;
 
 
-  template <int SequenceDims>
+  template <int32_t SequenceDims>
   static void Lookup(const std::string& name, std::optional<SyncMethod<SequenceDims>>& functor) {
     const auto iter = _METHOD_TABLE.find(name);
     if (iter != _METHOD_TABLE.end())
